Added lynx.pp
lynx.pp has now been added to ensure correct installation on RedHat based systems.
RedHat requires the yum config manager to be configured for this installation to proceed.
To enable this an exec statement is conditionally executed if lynx is not installed on a RedHat system.

Currently Debian is also supported.