Added csh.pp and fixed lynx.pp issue.
lynx.pp previoulsy had a require statement that was changed to a before keyword.
This was believed to be the correct method of running a command before install.
However this failed and has now been reverted back to a require statement as previous.

Also included a csh.pp to ensure puppet installs csh.