Modified vnc.pp to start managing vnc config setup (not yet complete).
Currently copies template file post installation to appropriate position.
Then providing the service is not running, meaning that the installation
was already successful, it edits the config file for root user.

Still need to reload the daemon, initialize the password, enable
the service and start/manage the service (including config subscription).