Created a mount resource in users.pp to mount my student folder to /home/becca/titan.
Manually mounting the folder works fine and the resource also manages
the creation of the folder, however the automated mounting fails
no doubt due to the authentication options not being in the resource.

Manually mounting prompts for a password and the initial attempt asks
to validate ECDSA keys.

Will still have to figure out how to automatically mount the file system.